requirements.txt
langchain==0.1.16
langchain-community==0.0.33
langchain-openai==0.1.4
python-dotenv==1.0.1
Flask==3.0.2
requests==2.32.5
selenium==4.20.0
